Modularised abstract class
Limiter abstract class seperate from Vertex based implementation for
easier reading.
